==> 01-manual-data-feat-eng.ipynb <==
# Load libraries

# pip install fsspec / conda install -c conda-forge fsspec
# pip install s3fs
import pandas as pd
import numpy as np
import boto3
import os
import math
from sagemaker import get_execution_role

role = get_execution_role()

# Load data from one file

bucket_name = 'eliezerraj-908671954593-dataset'
prefix_name = 'payment/glue-job'
file_name = 'run-1713493427814-part-r-00000'

data_location = 's3://{}/{}/{}'.format(bucket_name, prefix_name, file_name)

df = pd.read_csv(data_location)

print("---------------------------------")
print(f"Shape of dataframe '{df.shape}':")

pd.set_option('display.max_columns', 500)     # Make sure we can see all of the columns
pd.set_option('display.max_rows', 20)         # Keep the output on one page
df.head()

# Load data from a chunk of files in a folder

s3 = boto3.client('s3')

bucket_name = 'eliezerraj-908671954593-dataset'
prefix_name = 'payment/glue-job/'

# List objects in the specified bucket and prefix
response = s3.list_objects_v2(Bucket=bucket_name, Prefix=prefix_name)

# Extract the list of file keys from the response
file_keys = [obj['Key'] for obj in response.get('Contents', [])]

# Read each file
df = pd.DataFrame()

for obj_key in file_keys[1:]:
    # Read the object content
    obj_content = s3.get_object(Bucket=bucket_name, Key=obj_key)['Body'].read()
    print(f"Content of file '{obj_key}':")

    data_location = 's3://{}/{}'.format(bucket_name, obj_key)

    df_temp = pd.read_csv(data_location)
    df = pd.concat([df, df_temp], ignore_index=True)
    print(f"Shape of chunk '{df.shape}':")
    print("---------------------------------")

print("---------------------------------")
print(f"Shape of dataframe '{df.shape}':")


pd.set_option('display.max_columns', 500)     # Make sure we can see all of the columns
pd.set_option('display.max_rows', 20)         # Keep the output on one page
df.head(3)

# Featuring Engeniring

# Filter dataframe
df = df.filter(['fraud',
                'payment_at',
                'terminal_name',
                'coord_x',
                'coord_y',
                'card_type',
                'card_model',
                'mcc',
                'amount',
                'tx_1d',
                'avg_1d',
                'tx_7d',
                'avg_7d',
                'tx_30d',
                'avg_30d',
                'time_btw_cc_tx'], axis=1)

df.head(3)

# Create a new column distance
df.insert(5, "distance", 0)
for ind in df.index:
    p1 = [0, 0]
    p2 = [df['coord_x'][ind], df['coord_y'][ind]]
    distance = math.sqrt(((p1[0] - p2[0]) ** 2) + ((p1[1] - p2[1]) ** 2))
    df["distance"][ind] = distance

# extract hour
df["ts_payment_at"] = pd.to_datetime(df['payment_at'])
print(df.ts_payment_at.dtype)
df["hour"] = df.ts_payment_at.dt.hour

# 2nd filter dataframe
df = df.filter(['fraud',
                'hour',
                'distance',
                'card_type',
                'card_model',
                'mcc',
                'amount',
                'tx_1d',
                'avg_1d',
                'tx_7d',
                'avg_7d',
                'tx_30d',
                'avg_30d',
                'time_btw_cc_tx'], axis=1)

df.head(3)

df = pd.get_dummies(df, columns=['card_model', ], prefix='card_model', dtype=int, drop_first=False)
df = pd.get_dummies(df, columns=['card_type', ], prefix='card_type', dtype=int, drop_first=False)

df = df.filter(['fraud',
                'distance',
                'card_model_CHIP',
                'card_model_VIRTUAL',
                'card_type_CREDIT',
                'amount',
                'tx_1d',
                'avg_1d',
                'tx_7d',
                'avg_7d',
                'tx_30d',
                'avg_30d',
                'time_btw_cc_tx'], axis=1)

df.head(3)

# Split Data

train_df, validation_df, test_df= np.split(df.sample(frac=1, random_state=1729), [int(0.7 * len(df)), int(0.9 * len(df))])

print("---------------------------------")
print(f"Shape of dataframe '{train_df.shape}':")
print(f"Shape of dataframe '{validation_df.shape}':")
print(f"Shape of dataframe '{test_df.shape}':")
print("---------------------------------")

train_df.to_csv('train_data.csv', index=False, header=False)
validation_df.to_csv('validation_data.csv', index=False, header=False)
test_df.to_csv('test_data.csv', index=False, header=False)

bucket_name = 'eliezerraj-908671954593-dataset'
prefix_name_load_file = 'payment/notebook'

boto3.Session().resource('s3').Bucket(bucket_name).Object(os.path.join(prefix_name_load_file, 'train/train_data.csv')).upload_file('train_data.csv')
boto3.Session().resource('s3').Bucket(bucket_name).Object(os.path.join(prefix_name_load_file, 'validation/validation_data.csv')).upload_file('validation_data.csv')
boto3.Session().resource('s3').Bucket(bucket_name).Object(os.path.join(prefix_name_load_file, 'test/test_data.csv')).upload_file('test_data.csv')
