==> 01.1-sagemaker-data-feat-eng.ipynb <==
# Load libraries

# pip install fsspec / conda install -c conda-forge fsspec
# pip install s3fs
import pandas as pd
import numpy as np
import boto3
import os
from sagemaker import get_execution_role

role = get_execution_role()

# Load data from one file

bucket_name = 'eliezerraj-908671954593-dataset'
prefix_name = 'payment/glue-job'
file_name = 'run-1713493427814-part-r-00000'

data_location = 's3://{}/{}/{}'.format(bucket_name, prefix_name, file_name)

df = pd.read_csv(data_location)

print("---------------------------------")
print(f"Shape of dataframe '{df.shape}':")

pd.set_option('display.max_columns', 500)     # Make sure we can see all of the columns
pd.set_option('display.max_rows', 20)         # Keep the output on one page
df.head()

# Load data from a chunk of files in a folder

s3 = boto3.client('s3')

bucket_name = 'eliezerraj-908671954593-dataset'
prefix_name = 'payment/glue-job/'

# List objects in the specified bucket and prefix
response = s3.list_objects_v2(Bucket=bucket_name, Prefix=prefix_name)

# Extract the list of file keys from the response
file_keys = [obj['Key'] for obj in response.get('Contents', [])]

# Read each file
df = pd.DataFrame()

for obj_key in file_keys[1:]:
    # Read the object content
    obj_content = s3.get_object(Bucket=bucket_name, Key=obj_key)['Body'].read()
    print(f"Content of file '{obj_key}':")

    data_location = 's3://{}/{}'.format(bucket_name, obj_key)

    df_temp = pd.read_csv(data_location)
    df = pd.concat([df, df_temp], ignore_index=True)
    print(f"Shape of chunk '{df.shape}':")
    print("---------------------------------")

print("---------------------------------")
print(f"Shape of dataframe '{df.shape}':")


pd.set_option('display.max_columns', 500)     # Make sure we can see all of the columns
pd.set_option('display.max_rows', 20)         # Keep the output on one page
df.head(3)

df.to_csv('payment.csv', index=False, header=True)

# Upload
bucket_name = 'eliezerraj-908671954593-dataset'
prefix_name_load_file = 'payment/notebook'

boto3.Session().resource('s3').Bucket(bucket_name).Object(os.path.join(prefix_name_load_file, 'payment.csv')).upload_file('payment.csv')

# Featuring Engeniring

from sagemaker.sklearn.processing import SKLearnProcessor
from sagemaker.processing import ProcessingInput, ProcessingOutput
from sagemaker import get_execution_role

from sagemaker import Session
import sagemaker
from sagemaker import get_execution_role
from sagemaker.workflow.parameters import ( 
 ParameterInteger,
 ParameterString,
 ParameterFloat)

processing_instance_count = ParameterInteger(name="ProcessingInstanceCount", default_value=1)

framework_version = "1.0-1"
sklearn_processor = SKLearnProcessor(
    framework_version=framework_version,
    role=role,
    instance_type="ml.m5.large",
    instance_count=1,
    base_job_name='mlops-payment-feature-eng'
)

bucket_name = 'eliezerraj-908671954593-dataset'
prefix_name = 'payment/notebook'
file_name = 'payment.csv'

input_source = 's3://{}/{}/{}'.format(bucket_name, prefix_name, file_name)
data_location = 's3://{}/{}/{}'.format(bucket_name, prefix_name, 'requirements.txt')
boto3.Session().resource('s3').Bucket(bucket_name).Object(os.path.join(prefix_name_load_file, 'requirements.txt')).upload_file('requirements.txt')

print("---------------------------------")
print(f"bucket_name '{bucket_name}':")
print(f"prefix_name '{prefix_name}':")
print(f"input_source '{input_source}':")
print(f"data_location '{data_location}':")

processor_args = sklearn_processor.run(
    inputs=[
        ProcessingInput(source=input_source,
                        destination="/opt/ml/processing/input"),
        ProcessingInput(source=data_location,
                        destination="/opt/ml/processing/input/req/")
    ],
    outputs=[
        ProcessingOutput(output_name="train",
                         source="/opt/ml/processing/train",
                         destination=f"s3://{bucket_name}/{prefix_name}/output/train"),
        ProcessingOutput(output_name="validation",
                         source="/opt/ml/processing/validation",
                         destination=f"s3://{bucket_name}/{prefix_name}/output/validation"),
        ProcessingOutput(output_name="test",
                         source="/opt/ml/processing/test",
                         destination=f"s3://{bucket_name}/{prefix_name}/output/test")],
    code=f"feat-v02.py",
)
