==> 02-build-model.ipynb <==
# Load libraries

# pip3 install s3fs
from sagemaker import Session
import sagemaker
import boto3
import re
from sagemaker import get_execution_role
import numpy as np
import pandas as pd
import os

role = get_execution_role()

bucket_name = 'eliezerraj-908671954593-dataset'
prefix_name = 'payment/notebook/output'

train_path = f"s3://{bucket_name}/{prefix_name}/train"
validation_path = f"s3://{bucket_name}/{prefix_name}/validation"
test_path = f"s3://{bucket_name}/{prefix_name}/test"

x_train_data_file = f's3://{bucket_name}/{prefix_name}/train/train_data.csv'
x_train_input = sagemaker.inputs.TrainingInput(x_train_data_file, content_type='csv')
validation_data_file = f's3://{bucket_name}/{prefix_name}/validation/validation_data.csv'
validation_input = sagemaker.inputs.TrainingInput(validation_data_file, content_type='text/csv')

print("---------------------------------")
print(f"bucket_name '{bucket_name}':")
print(f"prefix_name '{prefix_name}':")

print(f"x_train_data_file '{x_train_data_file}':")
print(f"validation_data_file '{validation_data_file}':")

# Prepare the model

container = sagemaker.image_uris.retrieve(region=boto3.Session().region_name,
                                          framework='xgboost',
                                          version='latest')

sess = sagemaker.Session()

# algortimo class = 'binary:logistic'
hyperparameter_objective = 'reg:linear'

# round to run the training
hyperparameter_num_round = 100

# alpha - L1 regularization
hyperparameter_alpha = 5

# eta - step to prevent overfitting
hyperparameter_eta = 0.2

# deep size of trees
hyperparameter_max_depth = 5

xgb = sagemaker.estimator.Estimator(container,
                                    role,
                                    instance_count=1,
                                    instance_type='ml.m4.xlarge',
                                    output_path='s3://{}/{}'.format(bucket_name,
                                                                    prefix_name),
                                    sagemaker_session=sess)

xgb.set_hyperparameters(max_depth=hyperparameter_max_depth,
                        eta=hyperparameter_eta,
                        alpha=hyperparameter_alpha,
                        min_child_weight=6,
                        subsample=0.8,
                        silent=0,
                        objective=hyperparameter_objective,
                        num_round=hyperparameter_num_round)

xgb.fit({'train': x_train_input,
         'validation': validation_data_file})

# Model Validation

# Deploy Model
xgb_predictor = xgb.deploy(initial_instance_count=1,
                           instance_type='ml.m4.xlarge')

# Save model
xgb_predictor.serializer = sagemaker.serializers.CSVSerializer()

test_data_file = f's3://{bucket_name}/{prefix_name}/test/test_data.csv'
df_test_data = pd.read_csv(test_data_file, header=None)

print("---------------------------------")
print(f"df_test_data '{df_test_data.shape}':")

df_test_data_x = df_test_data.iloc[:, 1:]
df_test_data_y = df_test_data.iloc[:, :1]

df_test_data_x.head(3)

df_test_data_y.head(3)

def predict(data, predictor, rows=500 ):
    split_array = np.array_split(data, int(data.shape[0] / float(rows) + 1))
    predictions = ''
    for array in split_array:
        predictions = ','.join([predictions, predictor.predict(array).decode('utf-8')])

    return np.fromstring(predictions[1:], sep=',')

predictions = predict(df_test_data_x, xgb_predictor)

np.round(predictions)

pd.crosstab(index = df_test_data_y[0], 
            columns=np.round(predictions), 
            rownames=['actuals'], colnames=['predictions'])

# Check model evaluation

from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from math import sqrt

k = df_test_data_y.shape[1]
n = len(df_test_data_y)

RMSE = float(format(np.sqrt(mean_squared_error(df_test_data_y, predictions)),'.3f'))
MSE = mean_squared_error(df_test_data_y, predictions)
MAE = mean_absolute_error(df_test_data_y, predictions)
r2 = r2_score(df_test_data_y, predictions)
adj_r2 = 1-(1-r2)*(n-1)/(n-k-1)

print('RMSE =',RMSE, '\nMSE =',MSE, '\nMAE =',MAE, '\nR2 =', r2, '\nAdjusted R2 =', adj_r2) 

# Make a Prediction

payment_data = [{'distance': 100,
                 'card_model_CHIP': 0.,
                 'card_model_VIRTUAL': 1.,
                 'card_type_CREDIT': 1.,
                 'amount': 2000,
                 'tx_1d': 9,
                 'avg_1d': 1365.00,
                 'tx_7d': 17,
                 'avg_7d': 263.529412,
                 'tx_30d': 28,
                 'avg_30d': 238.714286,
                 'time_btw_cc_tx': 97582.0}]
df_predict = pd.DataFrame(payment_data)
df_predict

y_pred = xgb_predictor.predict(df_predict).decode('utf-8')
print(y_pred)
print(np.round(float(y_pred)))

import boto3
client = boto3.client(service_name="sagemaker")
runtime = boto3.client(service_name="sagemaker-runtime")
endpoint_name='xgboost-2024-04-22-00-43-07-882'

# Endpoint invocation
payload = b"100., 0., 1., 1., 400., 365.0, 17., 263.529412, 28., 238.714286, 97582.0"

response = runtime.invoke_endpoint(
    EndpointName=endpoint_name,
    Body=payload,
    ContentType="text/csv",
)

y_predict = response["Body"].read().decode()
print(y_predict)
print(np.round(float(y_predict)))

# Make Validation

from sagemaker.tuner import IntegerParameter, CategoricalParameter, ContinuousParameter, HyperparameterTuner

hyperparameter_ranges = {'eta': ContinuousParameter(0, 1),
                         'min_child_weight': ContinuousParameter(1, 10),
                         'alpha': ContinuousParameter(0, 5),
                         'max_depth': IntegerParameter(1, 10)}

objective_metric_name = 'validation:auc'

tuner = HyperparameterTuner(xgb,
                            objective_metric_name,
                            hyperparameter_ranges,
                            max_jobs=20,
                            max_parallel_jobs=3)

tuner.fit({'train': x_train_input, 'validation': validation_data_file})

boto3.client('sagemaker').describe_hyper_parameter_tuning_job(
HyperParameterTuningJobName=tuner.latest_tuning_job.job_name)['HyperParameterTuningJobStatus']

tuner.best_training_job()

tuner_predictor = tuner.deploy(initial_instance_count=1,
                           instance_type='ml.m4.xlarge')

tuner_predictor.serializer = sagemaker.serializers.CSVSerializer()

test_data_file = f's3://{bucket_name}/{prefix_name}/test/test_data.csv'
df_test_data = pd.read_csv(test_data_file, header=None)
df_test_data_x = df_test_data.iloc[:, 1:]
df_test_data_y = df_test_data.iloc[:, :1]

print("---------------------------------")
print(f"df_test_data '{df_test_data.shape}':")

tuner_predictions = predict(df_test_data_x, tuner_predictor)

pd.crosstab(index=df_test_data_y[0], columns=np.round(tuner_predictions), rownames=['actuals'], colnames=['predictions'])

from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from math import sqrt

k = df_test_data_y.shape[1]
n = len(df_test_data_y)

RMSE = float(format(np.sqrt(mean_squared_error(df_test_data_y, tuner_predictions)),'.3f'))
MSE = mean_squared_error(df_test_data_y, tuner_predictions)
MAE = mean_absolute_error(df_test_data_y, tuner_predictions)
r2 = r2_score(df_test_data_y, tuner_predictions)
adj_r2 = 1-(1-r2)*(n-1)/(n-k-1)

print('RMSE =',RMSE, '\nMSE =',MSE, '\nMAE =',MAE, '\nR2 =', r2, '\nAdjusted R2 =', adj_r2)

# Clean the endpoints

xgb_predictor.delete_endpoint(delete_endpoint_config=True)
tuner_predictor.delete_endpoint(delete_endpoint_config=True)
