==> 05-pipeline-fraud-model.ipynb <==
# Pipeline - Customer Classification (xboost) v.03.05.24

!pip install --upgrade boto3 fsspec s3fs

### Pipeline Begin

import boto3 
import pandas as pd 
import sagemaker
import os
from sagemaker.workflow.pipeline_context import PipelineSession 

pipeline_name = f"sagemaker-mlops-fraud-pipeline" 

sagemaker_session = sagemaker.Session()
region = sagemaker_session.boto_region_name 
role = sagemaker.get_execution_role()
pipeline_session = PipelineSession()
model_package_group_name = f"FraudModelPackageGroup"

# Load data from v1

#bucket_name = 'eliezerraj-908671954593-dataset'
#prefix_name = 'payment'
#file_name = 'payment.csv'

#input_dataset = 's3://{}/{}/{}'.format(bucket_name, prefix_name, file_name)
#input_requirement = 's3://{}/{}/{}'.format(bucket_name,prefix_name,'requirements.txt')
#model_path = f"s3://{bucket_name}/{prefix_name}/output"

#print("input_dataset : ",input_dataset)
#print("input_requirement : ",input_requirement)
#print("model_path : ",model_path)

#store_data = pd.read_csv(input_dataset)

#from sagemaker.workflow.parameters import ( 
#                         ParameterInteger, 
#                         ParameterString, 
#                         ParameterFloat) 

#auc_score_threshold = 0.75
#base_job_prefix = "fraud-model"

#processing_instance_count = ParameterInteger(name="ProcessingInstanceCount",
#                                             default_value=1)
#processing_instance_type = ParameterString(name="ProcessingInstanceType",
#                                           default_value="ml.m4.xlarge")
#training_instance_type = ParameterString(name="TrainingInstanceType",
#                                         default_value="ml.m4.xlarge")

#model_approval_status = ParameterString(name="ModelApprovalStatus",
#                                        default_value="PendingManualApproval")

# Load data from v2

s3_client = boto3.client('s3')

bucket_name = 'eliezerraj-908671954593-dataset'

prefix_name_dataset = 'payment/glue-job/'
prefix_name = 'payment'
file_name = 'payment.csv'

input_dataset = 's3://{}/{}/{}'.format(bucket_name, prefix_name, file_name)
input_requirement = 's3://{}/{}/{}'.format(bucket_name, prefix_name, 'requirements.txt')
model_path = f"s3://{bucket_name}/{prefix_name}/output"

print("---------------------------------")
print("1. input_dataset : ",input_dataset)
print("2. input_requirement : ",input_requirement)
print("3. model_path : ", model_path)

print("---------------------------------")
print("4. Create a dataset")

# List objects in the specified bucket and prefix
response = s3_client.list_objects_v2(Bucket=bucket_name, Prefix=prefix_name_dataset)

# Extract the list of file keys from the response
file_keys = [obj['Key'] for obj in response.get('Contents', [])]

# Read each file
df = pd.DataFrame()

for obj_key in file_keys[1:]:
    # Read the object content
    obj_content = s3_client.get_object(Bucket=bucket_name, Key=obj_key)['Body'].read()
    print(f"Content of file '{obj_key}':")

    data_location = 's3://{}/{}'.format(bucket_name, obj_key)

    df_temp = pd.read_csv(data_location)
    df = pd.concat([df, df_temp], ignore_index=True)
    print(f"Shape of chunk '{df.shape}':")
    print("---------------------------------")

print("---------------------------------")
print("5. Shape of dataframe :", df.shape)

print("---------------------------------")
print("6. Save the dataframe : ", input_dataset )

df.to_csv('payment.csv', index=False, header=True)
boto3.Session().resource('s3').Bucket(bucket_name).Object(os.path.join(prefix_name, 'payment.csv')).upload_file('payment.csv')

store_data = pd.read_csv(input_dataset)

from sagemaker.workflow.parameters import ( 
                         ParameterInteger, 
                         ParameterString, 
                         ParameterFloat) 

auc_score_threshold = 0.75
base_job_prefix = "fraud-model"

processing_instance_count = ParameterInteger(name="ProcessingInstanceCount",
                                             default_value=1)
processing_instance_type = ParameterString(name="ProcessingInstanceType",
                                           default_value="ml.m4.xlarge")
training_instance_type = ParameterString(name="TrainingInstanceType",
                                         default_value="ml.m4.xlarge")
model_approval_status = ParameterString(name="ModelApprovalStatus",
                                        default_value="PendingManualApproval")

### Step 1: Feature-engineering

# Define Processing Step for Feature Engineering
from sagemaker.sklearn.processing import SKLearnProcessor
from sagemaker.processing import ProcessingInput, ProcessingOutput
from sagemaker.workflow.steps import ProcessingStep

framework_version = "1.0-1"

sklearn_processor = SKLearnProcessor(
    framework_version=framework_version,
    instance_type="ml.m4.xlarge",
    instance_count=processing_instance_count,
    base_job_name="mlops-pipeline-fraud-feature-eng",
    role=role,
    sagemaker_session=pipeline_session,
)

processor_args = sklearn_processor.run(
    inputs=[
        ProcessingInput(source=input_dataset,
                        destination="/opt/ml/processing/input"),
        ProcessingInput(source=input_requirement,
                        destination="/opt/ml/processing/input/req/")
    ],
    outputs=[
        ProcessingOutput(output_name="train",
                         source="/opt/ml/processing/train",
                         destination=f"s3://{bucket_name}/{prefix_name}/output/train"),
        ProcessingOutput(output_name="validation",
                         source="/opt/ml/processing/validation",
                         destination=f"s3://{bucket_name}/{prefix_name}/output/validation"),
        ProcessingOutput(output_name="test",
                         source="/opt/ml/processing/test",
                         destination=f"s3://{bucket_name}/{prefix_name}/output/test")
    ],
    code=f"feature-engineering.py",
)

step_process = ProcessingStep(
    name="Featuring-Engineering-Fraud-Model",
    step_args=processor_args
)

print("step_process : ", step_process)

### Step 2: Define HyperParameter Tuning  Step

from sagemaker.estimator import Estimator
from sagemaker.inputs import TrainingInput
from sagemaker.tuner import (
    IntegerParameter,
    CategoricalParameter,
    ContinuousParameter,
    HyperparameterTuner,
)
from sagemaker.workflow.steps import TuningStep

image_uri = sagemaker.image_uris.retrieve(
                                            framework="xgboost",
                                            region=region,
                                            version="1.0-1",
                                            py_version="py3",
                                            instance_type=training_instance_type,
)
fixed_hyperparameters = {
                        "eval_metric": "auc",
                        "objective": "binary:logistic",
                        "num_round": "100",
                        "rate_drop": "0.3",
                        "tweedie_variance_power": "1.4"
}
xgb_train = Estimator(
                        image_uri=image_uri,
                        instance_type=training_instance_type,
                        instance_count=1,
                        hyperparameters=fixed_hyperparameters,
                        output_path=model_path,
                        base_job_name=f"fraud-train",
                        sagemaker_session=pipeline_session,
                        role=role,
)

hyperparameter_ranges = {
                        "eta": ContinuousParameter(0, 1),
                        "min_child_weight": ContinuousParameter(1, 10),
                        "alpha": ContinuousParameter(0, 2),
                        "max_depth": IntegerParameter(1, 10),
                        }

objective_metric_name = "validation:auc"

tuner = HyperparameterTuner(
    xgb_train,
    objective_metric_name,
    hyperparameter_ranges,
    max_jobs=2,
    max_parallel_jobs=2,
)

hpo_args = tuner.fit(
    inputs={
        "train": TrainingInput(
            s3_data=step_process.properties.ProcessingOutputConfig.Outputs["train"].S3Output.S3Uri, 
            content_type="text/csv",
        ),
        "validation": TrainingInput(
            s3_data=step_process.properties.ProcessingOutputConfig.Outputs["validation"].S3Output.S3Uri, 
            content_type="text/csv",
        ),
    }
)

step_tuning = TuningStep(name="HyperParameter-Tuning-Fraud-Model", step_args=hpo_args)

print("step_tuning : ", step_tuning)

### Step 3: Define the evaluation script and model evaluation step

# define model evaluation step to evaluate the trained model
from sagemaker.processing import ScriptProcessor
script_eval = ScriptProcessor(
    image_uri=image_uri,
    command=["python3"],
    instance_type=processing_instance_type,
    instance_count=1,
    base_job_name="script-fraud-eval",
    role=role,
    sagemaker_session=pipeline_session,
)

eval_args = script_eval.run(
     inputs=[
            ProcessingInput(
                source=step_tuning.get_top_model_s3_uri(top_k=0, s3_bucket=f"{bucket_name}/{prefix_name}", prefix="output"),
                destination="/opt/ml/processing/model"
            ),
            ProcessingInput(
                source=step_process.properties.ProcessingOutputConfig.Outputs["test"].S3Output.S3Uri,
                destination="/opt/ml/processing/test"
            )
        ],
    outputs=[
            ProcessingOutput(output_name="evaluation", 
                             source="/opt/ml/processing/evaluation",
                             destination=f"s3://{bucket_name}/{prefix_name}/output/evaluation"),
        ],
    code=f"evaluate-fraud.py",
)

from sagemaker.workflow.properties import PropertyFile

evaluation_report = PropertyFile(
    name="FraudModelEvaluationReport", output_name="evaluation", path="evaluation.json"
)

step_eval = ProcessingStep(
    name="Evaluation-Model-Fraud-Model",
    step_args=eval_args,
    property_files=[evaluation_report],
)

print("step_eval : ", step_eval)

### Step 4: Define a register model step

from sagemaker import Model
from sagemaker.workflow.model_step import ModelStep

model = Model(
    image_uri=image_uri,
    model_data=step_tuning.get_top_model_s3_uri(top_k=0,s3_bucket=f"{bucket_name}/{prefix_name}",prefix="output"),
    sagemaker_session=pipeline_session,
    role=role,
)

from sagemaker.model_metrics import MetricsSource, ModelMetrics
from time import gmtime, strftime

model_name = "xgboost-fraud-v3-" + strftime("%Y-%m-%d-%H-%M-%S", gmtime())

model_metrics = ModelMetrics(
    model_statistics=MetricsSource(
        s3_uri="{}/evaluation.json".format(
            step_eval.arguments["ProcessingOutputConfig"]["Outputs"][0]["S3Output"]["S3Uri"]
        ),
        content_type="application/json",
    )
)
register_args = model.register(
                                content_types=["text/csv"],
                                response_types=["text/csv"],
                                inference_instances=["ml.t2.medium", "ml.m5.large"],
                                transform_instances=["ml.m4.xlarge"],
                                model_package_group_name=model_package_group_name,
                                approval_status=model_approval_status,
                                model_metrics=model_metrics,
)

step_register = ModelStep(
    name="RegisterModel-Fraud-Model",
    step_args=register_args
)

print("step_register : ", step_register)

### Step 5: Define a condition step to check AUC score

from sagemaker.workflow.conditions import ConditionGreaterThan
from sagemaker.workflow.condition_step import ConditionStep
from sagemaker.workflow.functions import JsonGet
cond_lte = ConditionGreaterThan(
    left=JsonGet(
        step_name=step_eval.name,
        property_file=evaluation_report,
        json_path="classification_metrics.auc_score.value",
    ),
    right=auc_score_threshold,
)

step_cond = ConditionStep(
    name="CheckAUCScore-Fraud-Fraud",
    conditions=[cond_lte],
    if_steps=[step_register],
)

### Build and Trigger the pipeline run

import json
from sagemaker.workflow.pipeline import Pipeline

pipeline = Pipeline(
    name=pipeline_name,
    parameters=[
        processing_instance_count,
        processing_instance_type,
        training_instance_type,
        model_approval_status,
        input_dataset,
        auc_score_threshold,
    ],
    steps=[step_process,
           step_tuning,
           step_eval,
           step_cond],
)

definition = json.loads(pipeline.definition())

print(definition)

# Create a new or update existing Pipeline
pipeline.upsert(role_arn=role)
# start Pipeline execution
pipeline.start()
