==> 99-scaler.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

data = {'age': [25, 32, 47, 51, 28, 38, 45, 61],
        'gender': ['M', 'F', 'F', 'M', 'F', 'M', 'F', 'M'],
        'income': [50000, 55000, 60000, 65000, 48000, 58000, 62000, 68000],
        'country': ['US', 'UK', 'UK', 'US', 'US', 'UK', 'UK', 'US'],
        'target': [0, 1, 1, 0, 1, 1, 0, 0]}

df = pd.DataFrame(data)

df

df.info()

# Display the number of unique values in each column
unique_values_per_column = df.nunique()

print("Number of unique values in each column:")
print(unique_values_per_column)

#Initial data memory usage
memory_usage = df.memory_usage(deep=True)
total_memory_usage = memory_usage.sum()
print(f"\nTotal memory usage of the DataFrame: {total_memory_usage / (1024 ** 2):.2f} MB")

from sklearn.compose import make_column_transformer
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import cross_val_score

preprocessor = make_column_transformer(
    (StandardScaler(), ['age', 'income']),
    (OneHotEncoder(), ['gender', 'country'])
)

preprocessor

X_transformed = preprocessor.fit_transform(df)
X_transformed

from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, OneHotEncoder

preprocessor2 = ColumnTransformer(
 transformers=[
 ('standardscaler', StandardScaler(), ['age', 'income']),
 ('onehotencoder', OneHotEncoder(), ['gender'])
 ]
)

X_transformed2 = preprocessor2.fit_transform(df)
X_transformed2
